requirements.txt
asgiref==3.11.1
diff-match-patch==20241021
Django==5.2.11
django-import-export==4.4.0
django-modeltranslation==0.19.19
pillow==12.1.0
sqlparse==0.5.5
tablib==3.9.0
typing_extensions==4.15.0
tzdata==2025.3
